refactor: share application startup between both entry points
Startup was 180 lines inside `main()`. A second entry point — the Mortsom
test build — would have had to copy them, and a copy drifts: the two
builds would start the app differently, which is exactly what an
end-to-end harness must not have.

`bootstrapAndRun` now holds that sequence and both entry points call it.
It takes one argument, the relay seed list, which is the only thing a
build under test needs to do differently.

The static guard on subscribe-before-identity-init follows the code it
guards, and now covers both builds rather than the production one only.
